# Faster CD solver when `n_samples >> n_features`