The goal of this PR is to write a CD (BCD) solver when `n_samples >> n_features`.
Such configurations are solved much faster by pre-computing a Gram matrix XtX and updating the gradient (rather than the residuals) at every CD cycle.

A quick experiment with 1e6 samples and 300 features:
```
########
Lasso
########

Celer: 5.43s
Gram: 1.89s

##########
Group Lasso
##########

Celer: 43.41s
Gram: 4.23s
```