@mathurinm 2022-04-21
@PABannier thanks a lot ! I tried it and it really shines on the data by @sehoff

Caveat:  the stopping criteria are not the same. Monitoring the primal decrease is much looser than checking the duality gap. I adjusted the tolerance manually to obtain similar results