@sehoff 2022-04-22
Thanks a lot! I tried it out, and can confirm sizable speedups. I use warm-starts for celer, so I guess the figures are even biased towards celer.

**Note**: in the current setting of the gram-solver: ```res = gram_group_lasso(X, y, a, groups=grps, tol=1e-10, max_iter=10_000, check_freq=10)``` with very small alphas (alpha_max/1000), the solver does not convergence after ```max_iter```. 

Do you have any suggestion on a reasonable decrease in the tolerance? Because for ```tol=1e-9``` it reaches convergence after ~3500 iterations (the data is slightly different to the one I provided)
![speed_comparison_celer_gram](https://user-images.githubusercontent.com/34772392/164713239-0e65516b-b735-42e2-8643-b2c25efc6b57.jpg)
.