@PABannier 2022-04-24
Quick update:
- [ ] Gram CD
- [ ] Gram BCD
- [ ] Gram CD - FISTA
- [ ] Gram BCD - FISTA