@mathurinm 2022-04-25
@sehoff support for wieghts (inifnite ones too) is here. Let us know how it works !