@PABannier 2022-04-26
@sehoff For very small alphas, I'd recommend using FISTA instead. Have a look at the `gram_fista_group_lasso` function!