@sehoff 2022-04-26
> @sehoff support for weights (inifnite ones too) is here. Let us know how it works !
> @sehoff For very small alphas, I'd recommend using FISTA instead. Have a look at the gram_fista_group_lasso function!

Here are the results of my comparisons, where I used the data I provided in the dropbox, and a ```tol=1e-8``` in all solvers. Furthermore, the results I show refer to Case 3.1., i.e., with one weight set to infinity. As a side remark, this highlights that *all three* solvers can handle infinite weights! 
**Bottomline**: especially for (very) small alphas the Gram-based solvers outperform the one in celer significantly. So depending on the grid one eventually searches, each solver has its advantages.
**Comparison 1:**
![speed_comparison_celer_gram](https://user-images.githubusercontent.com/34772392/165371281-0ccec014-0201-4251-b5a2-13535b694189.jpg)
**Comparison 2:** *Note: I leave out celer here, because for alpha_max*1e-3 it takes quite long for convergence.*
![speed_comparison_bcd_fista](https://user-images.githubusercontent.com/34772392/165371278-4ffa31dd-1206-4474-b25e-a629b5a6b2eb.jpg)