@PABannier 2022-04-26
@sehoff indeed, Celer is particularly efficient in settings where `n_features >> n_samples`. It implements a working set strategy, that is particularly useful in high-regularization regime (where there are few active features). In your first figure, for low level of regularization, since your design matrix has way more samples than features, Celer working set strategy is less useful.

Besides, for your data, the Gram solver has a cheaper update since `X.T @ X` is of size (n_features, n_features) (X.T @ X being a useful quantity at every gradient update).