@PABannier 2022-10-22
@mathurinm do we want to keep this PR open? FISTA and Gram-solver have been merged