@mathurinm 2022-10-22
This one supports groups while the merged PR doesn't, so it does not hurt to leave it open IMO