@mathurinm 2024-04-11
note: this is a PR from the `skglm` repo, instead it should be done from a branch of your fork @PABannier